+++ src/Pin.h
@@ -19,6 +19,6 @@ public:
     void off(){}
     void toggle(){}
     void set(int status_){}
-    float analogReadp(){}
+    int analogReadp(){}
 };
 #endif
